feat: jest container service unit tests
This commits adds container service unit tests for all services provided.

To run tests you need to be in Backend and run 'yarn test'

+++ backend/src/services/Products/container.service.ts
@@ -10,6 +10,15 @@ export class ContainerService {
         description: string;
     
       }): Promise<any> {
+        if (UserInput.name === '') {
+            throw new Error('Name is required');
+        }
+        if (UserInput.material === '') {
+            throw new Error('Material is required');
+        }
+        if (UserInput.description === '') {
+            throw new Error('Description is required');
+        }
         try {
           const newContainer = await prisma.container.create({
             data: {
@@ -36,6 +45,9 @@ export class ContainerService {
 
     public async findByID(id: number): Promise<any | null> {
         try {
+            if (!id) {
+                throw new Error('ID is required');
+            }
             const container = await prisma.container.findUnique({
                 where: {id},
             });
@@ -48,6 +60,9 @@ export class ContainerService {
 
     public async deleteContainer(id: number): Promise<any | null>{
         try {
+            if(await this.findByID(id) === null){
+                throw new Error('Container not found');
+            }
             const container = await prisma.container.delete({
                 where: {
                     id: id
@@ -68,6 +83,15 @@ export class ContainerService {
             const containerExists = await this.findByID(id);
             if(containerExists === null){
                 throw new Error('Container not found');
+            }
+            if(UserInput.name === ''){
+                throw new Error('Name is required');
+            }
+            if(UserInput.material === ''){
+                throw new Error('Material is required');
+            }
+            if(UserInput.description === ''){
+                throw new Error('Description is required');
             }
             const container = await prisma.container.update({
                 where: {id},
